change in txn ordering type

+++ sovrin_common/txn_util.py
@@ -113,7 +113,7 @@ def getTxnOrderedFields():
         (f.IDENTIFIER.nm, (str, str)),
         (f.REQ_ID.nm, (str, int)),
         (f.SIG.nm, (str, str)),
-        (TXN_TIME, (str, int)),
+        (TXN_TIME, (str, float)),
         (TXN_TYPE, (str, str)),
         (TARGET_NYM, (str, str)),
         (VERKEY, (str, str)),
